+++ merkle/tree.go
@@ -1,9 +1,6 @@
 package merkle
 
 import (
-	"bytes"
-	"encoding/hex"
-	"fmt"
 	"github.com/spacemeshos/go-spacemesh/log"
 	"github.com/syndtr/goleveldb/leveldb"
 )
@@ -91,34 +88,3 @@ func NewTreeFromDb(rootHash []byte, userDataFileName string, treeDataFileName st
 	return mt, nil
 }
 
-func (mt *merkleTreeImp) CloseDataStores() error {
-
-	err := mt.treeData.Close()
-	if err != nil && err != leveldb.ErrClosed {
-		log.Error("Failed to close tree db %v", err)
-		return err
-	}
-
-	err = mt.userData.Close()
-	if err != nil && err != leveldb.ErrClosed {
-		log.Error("Failed to close user db %v", err)
-		return err
-	}
-
-	return nil
-}
-
-// Print the tree to a string
-func (mt *merkleTreeImp) Print() string {
-	buffer := bytes.Buffer{}
-	buffer.WriteString("\n------------\n")
-	if mt.root == nil {
-		buffer.WriteString("Merkle Tree: Empty tree.\n")
-	} else {
-
-		buffer.WriteString(fmt.Sprintf("Merkle tree: root hash <%s>\n", hex.EncodeToString(mt.GetRootHash())[:6]))
-		buffer.WriteString(mt.root.print(mt.userData, mt.treeData))
-	}
-	buffer.WriteString("------------\n")
-	return buffer.String()
-}
